pyproject.toml
[project]
name = "search-harvester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "cachetools>=5.5.2",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "prometheus-client>=0.21.1",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "readability-lxml>=0.8.4.1",
    "structlog>=25.3.0",
    "types-cachetools>=5.5.0.20240820",
    "uvicorn>=0.34.2",
]
